+++ railway-app-station/src/main/java/com/railway/station_service/SheduleItemRepository.java
@@ -2,7 +2,7 @@ package com.railway.station_service;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface SheduleItemRepository extends CrudRepository<SheduleItem, Integer> {
+public interface SheduleItemRepository extends CrudRepository<SheduleItem, Long> {
 
 	
 }
